+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 * Fixed HMR in `piral-cli-webpack5` when running `piral debug`
 * Fixed issue ignoring `--log-level` in the initializers
-* Fixed shared submodule dependencies in `peerDependencies` (#342)
+* Fixed shared submodule dependencies in `peerDependencies` (#341)
 * Updated to use React 17 (#312)
+* Updated `withApi` to use `context` instead of `converters`
 * Added `feed` parameter to `pilet debug` (#344)
 * Added `userStore` as an option for `piral-oidc` (#345)
+* Added `wrappers` to enable central component transformations
 
 ## 0.12.4 (November 24, 2020)